goldenease_backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.2
pydantic==2.11.7
typing_extensions==4.14.1

# Authentication and security
bcrypt==4.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.6
cryptography==45.0.5

# Database
SQLAlchemy==2.0.42

# PDF and document processing
PyMuPDF==1.26.3
pdfplumber==0.11.7
pillow==11.3.0

# OCR (Optical Character Recognition)
pytesseract==0.3.13

# Computer vision (headless version for server deployment)
opencv-python-headless==4.8.1.78
numpy==1.24.3

# HTTP requests
requests==2.31.0

# Cloud services (optional)
boto3==1.34.0